requirements.txt
fastapi
uvicorn
langchain
python-dotenv
openai 